Wall Jump update 2.0
Inverted the direction calculation that was there from the last update. Now launches when facing away from the wall and goes higher when facing towards the wall, instead of the other way around.

Also added a crude reset button that teleports player back to start location, if one were to fall off the map, for testing purposes.