Wall Jump Update 2.3
Altered how the wall info is cached, due to bug discovered by pear. It now will cache 100% of the time as long as the surface hit is closer than the previous.

Changed the way that the wall jump rotation calculation is performed. It is now based on the wall normal itself and not just the current (cached) velocity direction.