Wall Jump Update 3.1
Turned the Wall Jump functionality into an Actor Component to declutter the Proto Action Character class.

By doing this I created it's own interface (C++) on the chance we need to use it with another class.

All defaults for wall jumping will be found under the Wall Jump Component inside the BP.

Also brought over the powerup system that I created a few weeks ago. Includes some classes that we may or may not use, such as healing/speed boost powerups.

It's a modular system that uses 3 parts, the Pickup Actor (base location for powerup), Powerup Actor (the physical representation of the powerup), and the component that gets created onto the actor that overlaps the pickup actor).

By porting it over I also felt the need to create a Powerup Interface (C++) so that it can be used with other classes later on, if need be.

I did create a Double Jump Charge "powerup" with the system and placed one onto the test map for demonstration/testing purposes.

There are also a few materials that I created and am leaving in for use in the future, for the item pickups.

There was a lot added/changed in this, I hope I touched on everything.