Wall Jump Update 3.3
Wall Jump:
Altered the Double Jump mech to not fire until the jump button is released.
This also allowed the change to the Wall Jump, listed next.

You can now hold the jump key as you approach the wall to grab it, instead of getting lucky enough to hit it at the right time before you fall to your death.

Powerups:

Also a small update to the powerup classes because Unreal was not having a specific way that the replication was getting set inside the constructors.

Another change was changing to a dynamic setup to component overlaps, I believe this fixes the issue we were experiencing with the entire editor/game freezing momentarily when picking up powerups. (Hopefully)