Powerup Update 1.0.3
Updated the Powerup Base Component class to not call to destroy immediately on the same frame as when it was created. This was most likely the cause of the massive random error spike we were receiving  when overlapping with the powerup.

Was not tested extensively, but seemed to be the cause from the testing that was done.