Death Update 1.1
Added a basic form of the death mechanic.

Death/Respawn mechanic is about done inside of the character class.

Will need to update further with the level resetting upon respawning and then incorporate a checkpoint system that tracks progress and when someone respawns it will choose the most recent checkpoint instead of the beginning of the level.