Wall Jump Update 3.3.2
Gave the player a "Wall Hop" that allows the player to go straight up when looking towards the wall, when performing a wall jump. This force is editable in the BP.

Will have to figure out a method that doesn't allow the player to use the normal wall jump to go higher than this, by looking straight up.