Mine Update
Updated the mine class to have an "inner collision."

This is so that if we were ever to change the mine's static mesh, we could allow for a denser, more detailed mesh, and not suffer performance from collision calculations.

If static mesh is updated, will need to tweak the size of the inner collision to a size to appropriately encapsulate the static mesh.