Powerup Update 1.0.4
Revamped the Powerup system a tiny bit. Created a new class for "instant-use" powerup pickups, the way the other is designed, was mainly for long-term use/multiple ticks/long duration.

Because of this, there was an oversight to how actor components are handled in unreal, when creating and destroying in the same frame there can trigger a massive stutter. This update aims to prevent that by using a different class for powerups that can potentially cause that issue.

The new class is extremely simple, nothing much to note.

Re-did a bit of code in the pickup actor to accommodate for the "instant-use" powerups. Dynamic enough that it can recognize if you're using an "instant-use" or "normal" powerup class.

If by some means there is a class slotted in both the normal and instant-use slots, it will prefer the instant-use and completely ignore the normal one.