Powerup Update 1.0.6
Changed how the pickup actor handles the powerup actor, no longer destroys and respawns the object. This was in an attempt to completely remove the possibility that the player might pick up the object before it's fully spawned and cause a frame spike when it destroys itself.

Will have to eventually rework the old powerup system to accommodate for this change because of the actor component limitation that we seem to be running into.